refactor(parser): extract TensorLink types into dedicated header
