+++ src/parser.cc
@@ -6,26 +6,6 @@
 
 namespace Jetstream {
 
-bool TensorLink::resolved() const {
-    return producer.has_value() && !producer->module.empty() && !producer->port.empty() && !tensor.empty();
-}
-
-void TensorLink::requested(const std::string& block, const std::string& port) {
-    external = BlockEndpoint{block, port};
-    producer = std::nullopt;
-    tensor = Tensor{};
-}
-
-void TensorLink::produced(const std::string& module, const std::string& port, const Tensor& value) {
-    producer = ModuleEndpoint{module, port};
-    external = std::nullopt;
-    tensor = value;
-}
-
-void TensorLink::exposedAs(const std::string& block, const std::string& port) {
-    external = BlockEndpoint{block, port};
-}
-
 //
 // Helper Functions
 //
